Fix bug with standalone operators that explicitly use the operator keyword
Turns out, we should only set the thisTag if a function is known to be a method,
otherwise it will be an error.  We still need to ensure that the thisTag is set
to INTENT_TYPE for operator methods, so set it when the class adds a _this arg
if it didn't have one when the function was originally declared.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>