Adjust CG tests from U. Colorado collaboration
These CG tests were added in #17018.

The bulk of the changes is in updating to the currently-implemented syntax:
- remove `type` keyword in interface formals lists
- use `where implements IFC1(T1) && implements IFC2(T2)`
  instead of `where IFC(T)` or `where implements IFC1(T1), IFC2(T2)`

Other changes:
- added .notest to 1 test as I am unsure how or whether to test the desired
  feature, as it relates to unconstrained generics
- futurized a couple of tests because of unresolved or currently-leaning-
  the-opposite-direction features
- futurized a couple of tests because the compiler crashes
- removed a couple of duplicate tests

These tests remain NOTEST-ed, as the majority of them fail because they use
a required function with non-void return type. I suggest enabling testing
once this feature is implemented.

The remote second feature that is not implemented is generic implements
statement, ex. `R implements IFC;` where R is a generic record or a class.
Even if it is a class, this could be a generic declaration over memory
management and/or nilability.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>