chpldoc: display nil default values
This part of appendExpr() emits documentation for two sorts of cases:

var a: [something] int;
var o = nil;

When "something" in the first case is "", appendExpr() receives a
var->name == "nil".  It has code to emit this as "", to produce
"[] int" instead of "[nil] int".  This code was triggering in the
second case, leading to it being documented as "var o =".

The first case has printingType==true and the second case has
printingType==false, so use that to distinguish them.

Remove the future for this case in favor of adding an "= nil"
expression to three default value tests.

While I was in those tests, enable the 1-tuple cases which are now
working.

resolves #16941

---
Signed-off-by: Paul Cassella <gitgitgit@manetheren.bigw.org>