Improvements to automate running gpu tests
How to run on osprey (set up with Cray-CS and gasnet-ibv):
```
module load cudatoolkit
export CHPL_TEST_GPU=true
export CHPL_LAUNCHER_PARTITION=stormP100
start_test test/library/draft/cuBLAS/level1/
```

How to run on jupiter (set up with Cray-XC and ugni):
```
module load cudatoolkit
export CHPL_TEST_GPU=true
export CHPL_LAUNCHER_CONSTRAINT=K20XIV10
export CHPL_LAUNCHER_CORES_PER_LOCALE=20
start_test test/library/draft/cuBLAS/level1/
```

The main changes are to add a script and a PRECOMP to compile c_cublas.c
and to make the `-L` to the nvidia libraries less system specific by
basing the path off the location of `nvcc`. This works on jupiter and
osprey so it's somewhat portable, but it seems like there should be
something smarter we can do.

This also currently requires users to set `CHPL_TEST_GPU=true` in order
to run GPU tests, but ideally we would automatically run if `nvcc` or a
similar compiler was available and we were going to launch onto GPU
nodes. I'm not sure how to detect that we're going to launch onto a GPU
node though. For `CHPL_LAUNCHER=none` we can probably check for a GPU
directly, but I'm not sure if there's a reliable/programmatic way to
check with slurm if the nodes we're going to run on have GPUs.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>