Update LLVM tests to function with ASSERTS build
We configure bundled LLVM with -DLLVM_ENABLE_DUMP=ON
for asserts builds which I think is causing registers
to have names rather than just numbers.

This commit updates the tests to tolerate register names or numbers.
An alternative would be to use --ccflags -fdiscard-value-name to
request register names be discarded.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>