chpldoc: emit parentheses where needed
Emit parentheses in chpldoc output where needed to preserve the
semantics of the AST.  E.g., when the source-code expression we're
documenting is (a+b)*c), we need to emit parentheses.  But where the
expression is a+b+c, we don't, even though the structure of the AST is
the same (op (op a b) c).

Adding parentheses around every binary operator produces too many
parentheses: ((((a+b)+c)+d)+e), even when the source code expression
had none.  Instead, add parentheses only where the parenthesesless
expression would differ from the semantics of the AST.  e.g., AST:
(* a (+ b c)) needs to produce a*(b+c), not a*b+c.  This is mainly a
matter of operator precedence, plus certain operators not being
associative: (a-b)-c ==  a-b-c != a-(b-c), whereas (a+b)+c == a+(b+c).

When processing the "outer" operator -- the parent AST node -- pass
the operator string to the appendExpr() for the inner/child
expression(s) so that the child expressions can reason about their
precedence wrt the parent.

The processing of the parent could have done this reasoning, but it
would have to replicate the logic of appendExpr(Expr*) to
appendExpr(CallExpr*) to determining that the child is in fact an
operator.  Instead, the parent passes the operator string, and the
cases that don't care ignore it.

Add a test with a variety of operators and precedence patterns.

Resolves #16939

While I was here, add support for unary +, which was falling through
to the default function call case: var x = +5 was being rendered +(5).
Add a unary + case to a few tests.

---
Signed-off-by: Paul Cassella <gitgitgit@manetheren.bigw.org>