Remove -DCHPL_OPTIMIZE, -DCHPL_DEBUG, -DNDEBUG in the compiler
The compiler reads in a list of includes and defines to pass to the backend
compiler from a file. If that file contains any of the above defines, remove
them from the arg list.

Remove the code to strip them out when creating the file in the first place
so that other uses of it will still have them.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>