Fix wrong_error_message_april_26_2010
It suprised me that the top-level writeln doesn't return
anything - I made issue #17179 about that.

The difference in behavior in wrong_error_message_april_26_2010.chpl
is due to:
 1. writeln doesn't return anything
 2. cast argument order swapped, so that error is encountered first.

I added variants to make the situation clearer and more self contained.
In particular the variants should not change behavior if/when writeln
returns bool.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>