Add a temporary value in BytesStringCommon getCStr
An earlier commit migrated away from defining a local cast within the
function to just inlining its contents. However it's important to store
the value resulting from x.buff into a variable (otherwise we are trying
to caste the reference to the buffer).

The issue comes up with
  test/memory/figueroa/LeakedMemory6.chpl

or even with hello.chpl when compiled with --no-local --no-inline.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>