Refine the fabric selection code a bit more.
I'm just making one more pass over this stuff, because when I started
working with it to add the next capability set we want to be able to
support I found that I wasn't entirely happy with how things were going.
This may seem like over-polishing, but I expect fabric selection to be
an area of frequent maintenance as networks and providers come and go
and providers evolve their capabilities, and it's worth applying some
effort to make that maintenance as pleasant as possible.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>