Update a passing .future
This test needed a couple small updates, but was otherwise passing. It needed
to use a couple modules that were previously included by default, and add a
require statement for a .h file it was trying to use. With those updates it
passes both with and without llvm.

Remove the .future file, the .compopts that contained --llvm, and the .skipif
that skipped it in non-llvm configurations.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>