Fix lifetime constraints computation when type has operator methods
We were accidentally breaking lifetime constraint computation when the function
was a method on a type with comparison operator methods.  Scope resolve used to
try to insert a method token and this argument, making the call's base
expression into a partial call.  Unfortunately, lifetime computation was not
expecting the call to be a partial call or a method call and so was breaking.
Avoid that issue by not transforming calls in lifetime clauses into method calls

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>