Fix bug with operator call and mismatched `this` arg
If the function we're examining is an operator method and the call is already
adjusted to be a method call, we could get into a situation where we'd
accidentally reject an appropriate match because the `this` actual did not have
the same type as the `this` formal.  But we shouldn't care about the `this`
arg at all for operator calls, so in the case where we could end up in this
situation, just skip the `this` actual and formal.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>