Add support for parallel scans of replicated arrays
This adds the optional doiScan routine to the Replicated array
representation, with similar limitations as on default, local arrays:
1D only, and only for scan operators whose state and result types
match.  Similar to other operations on Replicated arrays, it only
operates on the local replicand.  The implementation merely forwards
the call to the DR implementation that represents the locale's local
data.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>