Update preFold's handling of == and != to adjust for when they are method calls
This only matters for when the arguments are type variables - function
resolution would normally replace such comparisons on type arguments with the
result instead of following resolution for the call.  However, in the case where
the == call is inside a method of a type with its own == definition, we think
that it might be a call to that method so insert the method token and this
arguments.  This prevented the preFold analysis from happening prior to this
change.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>