Revert initElts, add pragma "unsafe" to defaultInitDataOnPrivatize
Michael suggested that we do not use `initElts`, as with such light
testing for PrivateDist, we might not catch cases in e.g.,
assignment where `initElts` causes the types of two PrivateArrs
to differ.

Instead, attach `pragma "unsafe"` to the variable used to
default-initialize the array slot. This pragma will silence the
compiler errors.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>