Improve launcher detection, especially on Cray CS systems
This improves our `CHPL_LAUNCHER` autodetection to pick better defaults
for gasnet, especially on Cray CS systems. Previously, on cray-cs
systems we'd usually default to `slurm-srun`, but for gasnet we want a
`slurm-gasnetrun_*` launcher (or just gasnetrun_* if it's not a slurm
system, but that's the common case.) This improves that detection, but
will still default to `slurm-srun` for ofi or comm=none.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>