Update the compiler to fix some bugs with initializers and operator methods
This issue fixes two bugs: first, it allows types to define initializers that
call operators which it happens to have its own definition for; second, it stops
generating a default assignment operator for types that have an assignment
operator method.  The latter was causing problems for determining if the type
needed to define an overload for the assignment operator due to the presence of
an init= method or cast operator.

With these changes, both of the tests work

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>