Fix test for promoted call
The expected output for the test and the test's operations differed - as it
stood prior to this change, the test was expecting the results of the array
comparison to be combined but I don't think that is something that normally just
happens.  At least, when the call is promoted, the result ends up being an array
of bools indicating the comparison between the individual elements of the array.
I could have updated the expected output, but instead I felt like updating the
output to match expectations by reducing the result into a single bool

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>