Move delayed-PUT-vis bitmap from task private data to the tx contexts.
The bitmaps that indicate which nodes were the targets of PUts which
need visibility forcing have been stored in task-private data.  But we
only use delayed PUT visibility when we have bound transmit contexts,
which are assigned permanently to threads and thus tasks.  That being
the case, here move those bitmaps from task private data to the tx
contexts.  This removes overhead from both communication and task
creation and destruction.

While here, do a little bit of cleanup in the buffered-op support.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>