Fix bug with promoting standalone op from method on type with op method
When the call had more actuals than the best match has formals, we got into some
trouble with how the wrapper was set up.  Ensure that we don't look for too many
formals, that we always have the new field on PromotionInfo set when we're using
it, and that the wrapper function also has the operator flag so we can adjust
the final number of arguments appropriately if we need to.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>