Initial attempt to pad structures to make them fit alignment requirements
This attempts to add padding for fields in structures that are misaligned.
It builds a vector of expected alignments for each field, then later walks the
fields and checks their offsets. If any offsets don't match the expected
alignment, it adds an array of int8s before the misaligned field.

I was hoping I could do this in place in the already created StructType, but
StructType.setBody() issues an error if it already has a body when you call
it again. So this is not currently working.

I think I probably also have to update the GEPMap to increase all following
field indices to account for the added padding fields.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>