Add a simple test of promoting read on an array of atomics
I ran into an issue where I'd only broken a bunch of large tests like SSCA2 or
HPL because we didn't have a test for the simplified version of what I broke.
This was frustrating because those tests generated a lot of output so it was
difficult to see in the start_test output exactly what about them was broken and
categorize the failure appropriately.  Having a simplified version of the edge
case will allow a faster solution to the problem, as future developers won't
have to search through the code in this file to understand what broke for that
failure mode (if it breaks again)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>