Enable the "while-var" construct for while-loops
This follows the suit of #17047 and the design (1b) in #13639.

This reuses some code added in #17047.
Instead of lowering the while-var construct in normalize,
as #17047 did for if-var, the lowering is done during parsing.
This is because for a while-loop the generated AST contains
TWO copies of the conditional, rather than one.

The ASTs for while-loops that do not use the while-var
construct are unaffected by this change.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>