Always use amudprun launcher with gasnet-udp
gasnet-udp requires the amudprun launcher. Change our chplenv detection
so we always default to amudprun and error if a user tries to set it so
something else. Note that we also allow `CHPL_LAUNCHER=none` if they
want to manually launch.

Trying to use a different launcher with gasnet-udp would result in vague
errors like: `GASNet: Invalid number of nodes: -nl4` and this should
make the issue much more apparent.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>