chpldoc parens: review comments: unary op exprs, more tests, INT_FATAL()
We need to add parens around a unary-op expression when it's the
"inner" op of a binary op with higher precedence than the unary op.
e.g. (-1)**2 and (-1):uint need the parens.  Have the unary operators
add parens where needed.

Add test cases for that, for a few other unary op cases, and for a mix
of function calls and operators.

Have opToPrecedence() call INT_FATAL() when the operator isn't
recognized.

Use comparisons to astrs for the operators with astrs.

Remove a stray newline added by the previous commit.

---
Signed-off-by: Paul Cassella <gitgitgit@manetheren.bigw.org>