Limit when we auto-detect slurm-gasnetrun to cray-* systems
In 17302 I changed chplenv launcher detection to use slurm-gasnetrun_*
if salloc is found and we're using `CHPL_COMM=gasnet` with a supported
launcher, but that was triggering for cases like chapcs where we want to
run locally, but slurm happens to be available. For now just restrict
the slurm-gasnetrun_* to cray-cs systems.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>