Fix bug causing us to improperly resolve other conversions for operator methods
resolveAlsoConversions was expecting dtMethodToken to be the only potential
additional argument but when dtMethodToken is present, there's also a "this"
argument that should be skipped as well.  Because we were grabbing the wrong
"toType", we thought the function was generic and would skip analysing it
appropriately.

In the case of the test that showed this issue, this meant failing to generate
an error message about overloading assignment for classes when other error
messages were triggered due to standalone overloads.

Fix resolveAlsoConversions to skip both the methodToken and the this argument
instead of just the method token argument.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>