Fix some issues with these new tests that showed up last night
I did not think to run memleaks testing on these tests, and hadn't
realized that they hadn't freed their unmanaged classes.  Do so here.

There was a portability issue on linux32 due to casting a c_void_ptr
to a 64-bit integer.  I've rewritten the test to avoid the cast, which
I believe should resolve the issue (but in the interest of time am
relying on the nightly testing to run on linux32 rather than doing it
myself; it passes on linux64 and seems like it should resolve the
lines of code in question).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>