Update the modules primer for changes this release
The access of the field, primary method, and secondary method of this record is
still fine because we resolve such things by going to type's definition point,
not because we ignore the privacy and limitations of `use`/`import` statements.
Replace the descriptive text to reflect that.

Additionally, add a demonstration of accessing a tertiary method due to listing
the type in an `only` list.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>