Update the spec and specialMethods primer to refer to tertiary methods
Differentiates between secondary methods and methods defined on a type outside
of the scope where the type is defined.  We've been using this term recently,
seemed like something that should be documented.

Note that I did not update the references to secondary methods in the modules
section as those references are already being altered.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>