Update for new import/use behavior w.r.t. methods
In the "Method Calls" section, note the general availability of primary and
secondary methods, and that tertiary methods are only available based on uses
and imports of the modules which define them.

In the "Importing Modules" and "Using Modules" sections, remove descriptions
saying that limitations and privacy of use and import are ignored for methods
and instead emphasize that the availability of primary and secondary methods is
universal while describing how to impact the availability of tertiary methods.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>