Ensure nilability is reflected is "called as" messages
This ensures that the argument type is shown as nilable when appropriate
in the "called as" lines of error messages, for example:

    errors-called-as.chpl:8: In function 'showU1':
    errors-called-as.chpl:13: error: Scoped variable curr would outlive the value it is set to
    errors-called-as.chpl:11: note: consider scope of cur
      errors-called-as.chpl:19: called as showU1(name: string, list: owned clist?)

DETAILS

When compiling this test:

    test/classes/delete-free/lifetimes/errors-called-as.chpl

the compiler gives the above error in checkLifetimesAndNilDereferences()
during the callDestructors pass.

The last line of it ("called as ....") is, however, printed after the pass
completes and USR_STOP() is invoked, calling printCallstackForLastError().
Therefore this is after callDestructors runs convertClassTypesToCanonical(),
which eliminates nilability declarations from all class types. This makes
the argument type non-nilable and the message showing a non-nilable type
incorrectly.

To fix this, issue USR_STOP() as start of convertClassTypesToCanonical(),
so that the types remains unaffected, should they need to be printed out.

I have observed this issue only upon ``--devel``.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>