CG: Fix memory leaks with strings
This fixes memory leaks arising when CG functions pass around the values
of interface types and get instantiated with strings.

The fix is to ensure that FLAG_INSERT_AUTO_DESTROY is added on CG
variables when it would for a record.

We have discussed making CG types behave like records w.r.t. memory
management throughout resolution. This remains outside the scope
of this PR.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>