Avoid moving element onto this too early in implementation of add()
If we call `moveToValue()` and pass the moved value to the hashtable
lookup function instead of the original value (which may be remote),
we will get strange lifetime errors in the test
'.../Set/types/testNilableTuple.chpl'.

The solution is to use my original implementation, which might cause
some unnecessary comm, but works.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>