Update to use operator keyword and fix a table
Use operator keyword for the operator definitions instead of proc in the spec.
While there, I noticed the operator table was not rendering right online or
locally, so adjust it to work.  Add align to the list of operator names in both
places (and by in the syntax tree now that operator names are distinct from
other function names)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>