Restore some 'inline' markings to IO.chpl
This restores some of the 'inline' markings to IO.chpl that were
removed in PR #16887.  They were removed hoping that they would speed
up compilation, but turned out not to have that large of an impact, at
least for things we measured or in aggregate on our testing system.
However, they did have a somewhat adverse effect on one of our fasta
implementations, which is IO-heavy.

This PR restores some, but not all, of those 'inline's according to
the following thinking:

* restores them to lock() and unlock() because (a) for unlocked
  files, the amount of code in those routines is minimal, and
  these proved to be on the critical path in Elliot's experiments

* restored them for read() and write() because these are the
  powerhouses of I/O, and most of the heavy lifting is done
  in per-argument subroutines, so we don't expect them to hurt
  much.

* restored them for the _to*() overloads on enums because what little
  complexity was in those routines is a param conditional, so they're
  really no more complex than the others

Meanwhile, these were maintained:

* left them off of mark/offset because these don't seem deeply
  time-critical to me and already seem nontrivial

* left them off of _read_binary_internal/_write_binary_internal/
  _read_one_internal/_write_one_internal because these seem like
  precisely the type of routine that we'd want to stamp out once
  per type being read/written, and they're not obviously trivial.

* left them off of readbits and chpl_do_format for similar reasons.

Number of lines in generated code *.c for a --fast compile of fasta.chpl:

         caseyb/fasta   arkouda
- master:    31682       894627
- this:      31990       908974
- % increase   <1%         1.6%

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>