Fix cast operator methods; add tests of cast 'cleanup.cpp' error messages
Presumably due to the user-facing ':' operator and the 'operator'
keyword being developed simultaneously, 'operator :' in method form
hasn't worked because the check that it's a binary operator didn't
take the extra method arguments into consideration.  This led to the
confusing error in which a legal cast operator method would cause the
compiler to complain that it didn't have two arguments, when it
clearly did.

This PR fixes that by making the cast checks in cleanup.cpp cognizant
of whether the function is a method or not.  It also adds some tests
to (a) lock in these error messages (which didn't seem to be tested
previously?) and (b) lock in case showing that a cast operator method
works.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>