Update the special methods primer to be more accurate
This primer was out-of-date with the addition of operator methods and was
incorrect about requiring two arguments before that (because we can override
unary operators as well).  Remove the requirement on the number of arguments and
the limitation that it can't be defined as methods.  We'll wait to explicitly
describe how to overload it as a method until those are more stable

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>