Use domainDistIsLayout to be sure not to get side effects
I was seeing comms in chpl__coerceMove for domains
relating to this line

  param rhsIsLayout = rhs.dist._value.dsiIsLayout();

the rhs.dist._value is computed for side effects even
though dsiIsLayout returns param. So, instead,
use a helper function that returns param to hide
the side effects.

I don't see any improvement in the number of GETs for the example from
chapel-private issue 1789 but perhaps an earlier commit already removed
this case.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>