Adjust _newPrivatizedClassHelp to use outer variable rather than arg
const privatizeData was passed as a const ref arg to
_newPrivatizedClassHelp but that prevents some RVF. Here, change it to
use an outer variable, which allows the compiler to figure out it is
immutable. I initially tried changing the intent to 'const in' but that
caused core dumps for some reason.

Brings GETs from 972 to 948 for the example from chapel-private issue 1789

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>