Replace "initializer" in user's guide with "initial value" where appropriate
I happened to stumble across this when looking at another thing.  I think this
user's guide entry was written before initializers were a key part of the
language.  Regardless, "initializer" now means a particular type of method on a
record or class, so instead use "initial value" when talking about that portion
of a variable declaration.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>