CG: Fix "not implemented" error with records
This removes the "interface is not implemented" error
when the interface declaration uses record types
instantiated with interface type(s).

To do this, the compiler pre-computes the instantiations
of records using implementing types that correspond to
the instantiations using interface type(s).
For example, if a required function in an interface
has a formal of the type `R(Self)`, where Self is the interface's formal,
and we are checking the interface's implementation for `int`,
we calculate the type `R(int)`.
These instantiations need to be substituted in several places
throughout resolution.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>