CG: Issue user errors upon CG innerCopy
Currently copyInner is not implemented properly
for InterfaceSymbol and ImplementsStmt because
the existing implementations do not handle
the non-AST data structures that these nodes store.
To prevent user surprises, issue user errors
if these methods are invoked.

None of the existing tests bump into these errors.

I am not adding any tests for these error messages
because this is not a desired behavior that we want to preserve.

Discussed with Michael.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>