Move 'align' and 'by' before the other operators to avoid weird rst highlighting
For some reason these keywords were getting italicized by Github, despite none
of the other 'words' encountering the same behavior.  It seems to have something
to do with following the list of other operators, as moving them before it fixes
the issue.

This did not affect the Sphinx build, but seemed worth resolving regardless.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>