Don't discard command output until we know we don't need it.
We got 'IndexError: string index out of range' from python in
run_command() because we tried to print what the target command had
produced to stderr, but our caller had told us not to collect that.
The details were: CHPL_HOST_PLATFORM was set to hpe-cray-ex and
CHPL_LIBFABRIC was set to system, but there was no libfabric module
loaded. Thus the 'pkg-config --exists libfabric' command run from
pkgconfig_get_compile_args() for a 'printchplenv --internal' failed,
and then we tried to print stderr.  But we hadn't asked for it.

Here, adjust: always collect stdout and stderr but only return to the
user the one(s) they asked for.  (I started to just collect stderr
always, since that's the only one run_command() itself might need,
but that actually ends up creating even more control flow.)

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>