First cut at support for GETs in message-order-fence MCM mode.
This is the first cut at support for GETs in the new MCM mode that uses
message orderings plus fences.  It's similar to the changes made to do
the same thing for PUTs.

While here I also replaced the "last op" variable in the tx context
table entries with a bitmap of nodes to which we've done an atomic op
but haven't guaranteed its effects are visible, since we need to track
this for each remote endpoint individually.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>