Add a test locking in another promotion bug with operator methods
This test is a distillation of a bug I encountered in a more complex program.
Basically, when using an operator on a type that has defined a standalone
operator from within a method on a type that has that same operator defined as
an operator method, if only one of the arguments were promoted we might fail to
include the remaining argument in the promotion wrapper.  Thankfully, the fix is
fairly simple

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>