Fix the promotion bug with types that define operator methods
We were accidentally failing to add non-promoted arguments to promotion wrappers
when inside a method on a type that happened to define its own operator overload
for the called operator as a method.  This was because for_vector will not
iterate over a vector that starts with a NULL entry, and we used for_vector when
determining the arguments for the promotion wrapper.

With this change, the original motivating case and the test case pass

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>