Vanquish operand1/operand2 vs. operand/comparand confusion in ofi.
I long ago adopted operand1/operand2 nomenclature in comm=ofi for the
two values involved in a CSWAP AMO.  This was a mistake.  (It was
derived from the ugni comm layer, and it was a mistake there, too.)  It
doesn't sufficiently distinguish the two values' purposes and in adds to
the confusion already present as a result of the Chapel module code,
runtime interface, and C standard atomics using arguments named like
"expected" and "desired" (in that order) for the value to compare to and
the value to replace it with, while the libfabric atomics interface has
these in the opposite order and uses "buf" (ugh) for the replacement
value and "compare" for the comparison value.  I also used "operand" in
some spots and "opnd" in others.  Here, name the new value "opnd" and
the comparison value "cmpr" everywhere, and maintain them in that order
everywhere inside the comm layer.  Do the required rename-and-reverse
from Chapel's (expected, desired) style to ofi's (operand, comparand)
style exactly once, in the chpl_comm_atomic_cmpxchg_*() interface
functions.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>