Re-enable knucleotide3
This test has been disabled for nearly a year now, though the version
at the CLBG site has continued to function (and is more up-to-date
than this one in that it applies a 'with' clause to a parallel
loop... not quite sure how they got out of sync without also
re-enabling this test).  Here, I'm updating the test to use explicit
reads/writes of syncs now that the implicit reads/writes are
deprecated.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>