Offsets can't be safely cast to 'int'; offset differences can
I'm not super-convinced that this is correct, but it may be not much
less correct than before.  I'm increasingly wondering whether fields
within arrays (like blk or factoredOffs) should either be `uint`s
always or perhaps be ints vs. uints depending on the idxType.  This
deserves more thought before merging, but in my sprint to get things
working, I think it's a step forward.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>