Improvements
* Store AST nodes within an chpl::ast namespace

The reason for it is that we will (at least in the short term)
have the old AST types available at the same time in some parts
of the compiler. We can store the old AST types in chpl::oast
and then use the namespaces to distinguish.

* Use an ASTContext for the uniqued strings

We should store any "global" information into ASTContext instead of using
actually global variables. That will allow more use cases of the library.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>