Enable "In module MMM:" error annotation
Print a line saying `In module MMM:` when the compiler prints
an error message for a module-level code, analogously to printing
`In function FFF:` for an error in a function.

Such a line is already printed in the developer mode, however it used
to be suppressed upon --no-devel. This PR enables it for the latter
as well.

This PR chooses to suppress it, however, in "trivial" cases, where
it looks more like noise than like adding value. The "trial cases" are:

* The name of the module is the same as the name of the file where
  the error occurred. There are ~1,200 tests in our test suite where
  this happens.

* The module starts on Line 1 of the file. There are ~12 cases for this.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>