Improve updateDatFiles.py's date range removal
Previously, updateDatFile.py required the start and end date to exist in
the file, but this was cumbersome when you just wanted to remove all
data in a range and some dat files might not have data for that range.

Now we just find the first date greater than the start date, and the
last date smaller than the end date and use that range instead of
requiring the specific start/end date being in the file.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>