Update halts to be compilerErrors for OOB cases and update tests
Because orderToEnum can result in an OOB error, the halt() calls
need to be converted to compilerError() calls in the param case.
Here, I updated existing tests that check for OOB cases at
execution time to use non-param values to preserve the behavior
and also added a new test for the param case to lock in the
new compile-time error.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>