Merge pull request #17504 from bradcray/open-param-enum-ranges
Add support for param for loops over open enum ranges

[reviewed by @dlongnecke-cray]

This permits my two recent efforts (support for param for loops over enum ranges; support for enum ranges) to work together.  The main step that was missing was to create a `param` version of the "order-to-enum" functions generated by the compiler for each enum (which also required conditionally changing the OOB-style halt() in the normal version into a compilerError() in the param version.  In addition, I had to (or "got to"!) re-generalize the utility routines that support open interval ranges to stop excluding enums now that they work for param cases as well.  Unfuturized a test to lock in the behavior, added a new test of the compilerError() and updated a few pre-existing tests to have them continue exercising the execution-time error (not their main reason for existing).